fix: Rename file to useDashboard
