+++ apps/dashboard/src/forms/ir_additional_info/ir_company_info.page.tsx
@@ -9,7 +9,7 @@ import {
     FormMessage
 } from "@/components/ui/form"
 import { Input } from "@/components/ui/input"
-import { useDashboard } from "@/shared/hooks/useRegistration"
+import { useDashboard } from "@/shared/hooks/useDashboard"
 import { HOST } from "@/shared/vars"
 import { asOptionalField } from "@/utils/zod_optional_field"
 import { zodResolver } from "@hookform/resolvers/zod"
